test-mhfixmsg: pwd(1)'s `-P' needed to avoid failure with symlinks.
Steven Winikoff reported to nmh-workers that one of this file's tests
failed in 1.7.  It attempts to manipulate the output of mhpath(1)
assuming it starts with the result of `pwd', but symlinks in the latter
may cause this to fail.  Use POSIX's `-P' option to ensure no symlinks,
as done in some other tests.