mhfixmsg.man: Replace UTF-8 `§' with `\(sc' troff escape.
man page source is ASCII.  GNU groff's troff's source is ISO 8859-1.
The `§' in UTF-8 looks like `Â§' in ISO 8859-1.  If those two bytes make
it to a UTF-8 terminal then it appears as `§', but target something
else, e.g. PDF, and the flaw shows.

Also adjust test-manpages to grep for bytes other than HT, and SPACE to
tilde, in the built man pages.  A test of the source could be also be
done for the Makefile's `all' target, but this one should remain to spot
the build process breaking things, e.g. non-ASCII in $prefix.