mhlsbr.c: Use variable for strncpy(3)'s size, not sizeof.
Silences gcc 8.1.1's warning that the size of the source was being used,
not the destination.  The destination has just been allocated to be the
size of the source so no overflow can occur.  Move existing variable
that later holds the buffer size to before the allocation and strncpy so
it can be used for those too.