Cleanup MIME header handling in mhbuild
Per discussion on nmh-workers, cleanup the handling of MIME headers
in mhbuild.  The following logic should now apply.

If mhbuild is run with -auto AND a MIME-Version header is detected, then
exit silently with no error (and leave the draft untouched).  Otherwise,
if a MIME-Version header or ANY Content-* header is found, exit with an
error.  Note that this means if mhbuild is run with -auto then an error
will be returned if any Content-* headers are seen.

Previously, mhbuild would silently eat any Content-Type headers seen
in an existing draft; this behavior was brought forward from MH 6 and
no one could explain the reason for this behavior.  This behavior has
been removed.