Remove final traces of TMA support
TMA (the TTI Trusted Mail Agent) was an early implementation of
encrypted mail and is long gone.  There were a few vestigial bits
of this support left in nmh, and this is finally removing them.
This includes the check for the "Encrypted" header field in various
scan formats and some code to return the code SCNENC from scan()
function where it would have special handling by programs like inc(1).